back-end/covid/processa/rt_predictor/requirements.txt
absl-py==0.13.0
appdirs==1.4.4
astunparse==1.6.3
backcall==0.2.0
Brotli==1.0.9
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.0
colorama==0.4.4
dash==1.20.0
dash-bootstrap-components==0.12.2
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==5.0.7
distlib==0.3.1
filelock==3.0.12
Flask==2.0.0
Flask-Compress==1.9.0
flatbuffers==1.12
furl==2.1.2
future==0.18.2
gast==0.4.0
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==2.0.0
jedi==0.18.0
Jinja2==3.0.0
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
orderedmultidict==1.0.1
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
plotly==4.14.3
prompt-toolkit==3.0.18
protobuf==3.17.3
psycopg2-binary==2.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
#pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.3
setproctitle==1.2.2
six==1.15.0
SQLAlchemy==1.4.15
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
ua-parser==0.10.0
Unidecode==1.2.0
urllib3==1.25.11
virtualenv==20.4.3
wcwidth==0.2.5
Werkzeug==2.0.0
wrapt==1.12.1
